finalize chat app

+++ src/components/LoginForm.js
@@ -1,39 +1,60 @@
 import React, {useState} from 'react';
 import axios from 'axios';
+import {Link, useHistory} from 'react-router-dom';
 
 const LoginForm = () => {
     const [username, setUsername] = useState('');
     const [password, setPassword] = useState('');
 
-    const [error, setError] = useState('')
+    const [error, setError] = useState('');
+    const history = useHistory();
 
     const handleSubmit = async (e) => {
         e.preventDefault();
-
+        if (!(username.length > 0 && password.length > 0)) return;
+        
         // authenticating
         const authObject = {'Project-ID': "07b488b1-b00e-42ca-88a1-52ffe93814aa",
                             'User-Name': username,
                             'User-Secret': password};
 
         try {
-            // getting messages from chatengine
-            await axios.get('https://api.chatengine.io/chats', {headers: authObject})
+            // getting users from chatengine
+            const users = await axios.get('https://api.chatengine.io/users/', 
+            {headers: {'PRIVATE-KEY': '{{c5a99d40-df00-41b2-91ec-6129f01306c3}}'}});
 
-            // passing username and password into localStorage
-            localStorage.setItem('username', username);
-            localStorage.setItem('password', password);
+            const  userExist = await axios.get('https://api.chatengine.io/users/me', 
+            {headers: authObject})
 
-            window.location.reload();
+            // users.data.map(user => console.log(user))
+            // mapping through the user array from chat engine to check if user exist
+            users.data.map((user) => {
+                // if user exists and password and username match
+                if (userExist.status === 200) {
+                    // getting chats from chat engine
+                    axios.get('https://api.chatengine.io/chats', {headers: authObject})
+                    // checking to see if username exist in localStorage
+                    if(localStorage.getItem('username') && localStorage.getItem('username')===username) {
+                        history.push('/chats');
+                    }
+                    else {
+                    // clearing localStorage
+                    localStorage.clear();
+                    // passing username and password into localStorage if user exists
+                    localStorage.setItem('username', username);
+                    localStorage.setItem('password', password);
+                    history.push('/chats')}
+                }
+            })
 
-        } catch (err) {
-            if(username !== authObject['User-Name'] || password !== authObject['User-Secret']) {
-                setError(`Oops,  the following error occurred: error in login credentials`)
+        } 
+        catch (err) {
+            if (err.message === 'Request failed with status code 403') {
+                setError(`Oops, Invalid username or password, if you do not have an account kindly register`)
             }
-            else {
-                setError(`Oops,  the following error occurred: ${err}`)
             }
-        }
     }
+
     return (
         <div className='wrapper'>
             <div className='form'>
@@ -63,6 +84,13 @@ const LoginForm = () => {
 
                     {/* error div */}
                     <h2 className='error'> {error}</h2>
+
+                    <div align="center">
+                        <button type='submit' className='button'>
+                            <Link to='/register'><span> Register</span></Link>
+                        </button>
+                    </div>
+                    
                 </form>
             </div>
         </div>
